feat: go_routerを使用したルーティングの実装
- アプリケーションのエントリーポイントを更新し、MaterialApp.routerを使用してgo_routerによるルーティングを設定しました。
- 新たにStartPageとResultPageを作成し、各ページへのルートをappRouterで管理するようにしました。
- 不要なresult_page.dartファイルを削除しました。
