+++ lib/app/app_router/app_router.dart
@@ -1,0 +1,22 @@
+import 'package:go_router/go_router.dart';
+import 'package:presen_neta/features/result/presentation/page/result_page.dart';
+import 'package:presen_neta/features/start/presentation/page/start_page.dart';
+
+/// アプリ全体のルーティングを管理するクラス。
+///
+/// [go_router] を用いて、各ページへのルートを集中管理する。
+
+final GoRouter appRouter = GoRouter(
+  routes: [
+    GoRoute(
+      path: '/',
+      name: 'start',
+      builder: (context, state) => const StartPage(),
+    ),
+    GoRoute(
+      path: '/result',
+      name: 'result',
+      builder: (context, state) => const ResultPage(),
+    ),
+  ],
+);
